en proceso arqueo

+++ src/main/java/net/resultadofinal/micomercial/model/wrap/CompraVentaWrap.java
@@ -1,18 +1,20 @@
 package net.resultadofinal.micomercial.model.wrap;
 
+import java.math.BigDecimal;
+
 public class CompraVentaWrap {
 	private String lista;
-	private Float total;
+	private BigDecimal total;
 	public String getLista() {
 		return lista;
 	}
 	public void setLista(String lista) {
 		this.lista = lista;
 	}
-	public Float getTotal() {
+	public BigDecimal getTotal() {
 		return total;
 	}
-	public void setTotal(Float total) {
+	public void setTotal(BigDecimal total) {
 		this.total = total;
 	}
 }
